Split manifest mapping loader
